staff delete fix

+++ LIMS/frontend/src/pages/AllStaff.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-
+import Swal from 'sweetalert2';
 import { useAuthContext } from "../hooks/useAuthContext";
 import $ from "jquery";
 
@@ -42,26 +42,51 @@ const AllStaff = () =>  {
 
     
 
-    const handleClick = async (id) => {
+    
+    const handleClickDelete = async (id) => {
         if(!user){
             return
         }
-        const response = await fetch('api/Staff/' + id,{
-            method: 'DELETE',
-            headers: {
-                'Authorization': `Bearer ${user.token}`
-            }
-        })
-        
-
-        if(response.ok)
-        {
+        const confirmed = await Swal.fire({
+          title: 'Are you sure?',
+          text: "You won't be able to revert this!",
+          icon: 'warning',
+          showCancelButton: true,
+          confirmButtonText: 'Yes, delete it!',
+          customClass:'alerts'
+        });
+      
+        if (confirmed.isConfirmed) {
+            const response = await fetch('api/Staff/' + id,{
+                method: 'DELETE',
+                headers: {
+                    'Authorization': `Bearer ${user.token}`
+                }
+            })
+          const json = await response.json();
+    
+      
+          if (response.ok) {
             const table = $('#example').DataTable();
             const row = table.rows(`[data-id ="${id}"]`);
             row.remove().draw();
-        }
-        
-    }
+    
+            Swal.fire(
+              {
+                title: 'Success',
+                text: 'Record has been deleted',
+                icon: 'success',
+                showConfirmButton: false,
+                timer: 2000,
+                timerProgressBar: true
+                
+            }
+            )
+            //dispatch({ type: "DELETE_EXPENSES", payload: json });
+          }
+        }
+    
+      };
 
     return (
         <div className="container">
@@ -82,7 +107,7 @@ const AllStaff = () =>  {
                          <td>{staff.name}</td>
                         <td>{staff.Eid}</td>
                         <td>{staff.post}</td>
-                        <td><button className="btnDelete" onClick={ () => handleClick(staff._id) }>delete</button></td>
+                        <td><button className="btnDelete" onClick={ () => handleClickDelete(staff._id) }>delete</button></td>
                     </tr>
                 ))}
                 </tbody>
